backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
Django==5.2
django-channels==0.7.0
django-cors-headers==4.7.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
